requirements.txt
boto==2.39.0
chardet==2.3.0
Collectfast==0.2.3
diff-match-patch==20121119
Django==1.9.2
django-countries==3.4.1
django-environ==0.4.0
django-filter==0.12.0
django-import-export==0.4.2
django-modeltranslation==0.11
django-ordered-model==1.1.0
django-solo==1.1.2
django-storages-redux==1.3.2
django-taggit==0.18.0
django-taggit-serializer==0.1.5
django-uuslug==1.1.7
djangorestframework==3.3.2
greenlet==0.4.9
gunicorn==19.4.5
Markdown==2.6.5
msgpack-python==0.4.7
neovim==0.1.0
Pillow==3.1.1
psycopg2==2.6.1
python-dateutil==2.4.2
python-slugify==1.2.0
pytz==2015.7
six==1.10.0
tablib==0.11.2
Unidecode==0.4.19
wheel==0.24.0
